+++ src/combinators/append.js
@@ -4,9 +4,9 @@ import Signal from '../Signal'
 import concat from './concat'
 
 /**
- * Emits a value `a` after the signal `s` has completed.
+ * Emits the values from an array `as` after the signal `s` has completed.
  *
- * @param a The value to append.
+ * @param as The value to append.
  * @param s {Signal} The signal.
  * @returns {Signal} A new signal.
  * @example
@@ -14,12 +14,12 @@ import concat from './concat'
  * import { Signal, append } from 'bulb'
  *
  * const s = Signal.fromArray[1, 2, 3]
- * const t = append(4, s)
+ * const t = append([4, 5, 6], s)
  *
- * t.subscribe(console.log) // 1, 2, 3, 4
+ * t.subscribe(console.log) // 1, 2, 3, 4, 5, 6
  */
-export function append (a, s) {
-  return concat(s, Signal.of(a))
+export function append (as, s) {
+  return concat(s, Signal.fromArray(as))
 }
 
 export default curry(append)